# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aegis Link API - URL Reputation Management",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/health/": {
      "get": {
        "operationId": "health_check",
        "tags": [
          "system"
        ],
        "description": "Health Check 엔드포인트\n\nDocker의 healthcheck에서 사용됩니다.\n서버가 정상적으로 실행 중인지 확인하는 간단한 엔드포인트입니다.",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/url-reputations/": {
      "get": {
        "operationId": "get_url_reputation",
        "tags": [
          "url-reputations"
        ],
        "description": "URL 평판 정보 조회\n\n사용자가 URL 클릭 시 해당 URL의 안전성을 확인하기 위해 사용합니다.\nDB에 해당 URL이 없으면 404를 반환합니다.",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UrlReputationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "create_url_reputation",
        "tags": [
          "url-reputations"
        ],
        "description": "URL 평판 정보 등록/업데이트\n\nAI Agent가 수집한 URL 정보를 DB에 저장합니다.\n이미 존재하는 URL이면 업데이트하고, 없으면 새로 생성합니다.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUrlReputation"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UrlReputationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateUrlReputation": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "nullable": true
          },
          "score": {
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/UrlStatus"
          },
          "url": {
            "type": "string"
          }
        },
        "required": [
          "url",
          "score",
          "status"
        ]
      },
      "GetUrlQuery": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          }
        },
        "required": [
          "url"
        ]
      },
      "UrlReputationResponse": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "nullable": true
          },
          "score": {
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/UrlStatus"
          },
          "url": {
            "type": "string"
          }
        },
        "required": [
          "url",
          "score",
          "status"
        ]
      },
      "UrlStatus": {
        "type": "string",
        "enum": [
          "safe",
          "warning",
          "block"
        ]
      }
    }
  },
  "tags": [
    {
      "name": "system"
    },
    {
      "name": "url-reputations"
    }
  ]
}